Rebate Oracle is a Certificate of Authority (PoAwR consensus specific)
PoAwR modifies PoA by adding a rebate reflux particle accelerator to the flux capacitor so where as if my calculations are correct, when this baby hits 88 miles per hour... you're gonna see some serious shit. No for real though this Rebate Oracle accepts the rebates in a decentralized manner. So as to predict the future DAO implementation, creating a whole new ball game folks! Decentralied DAO hopping could occur, let's hash it!

This implementation of the Rebate Oracle is a Certificate of Authority 
- guardian of the future polls
- creating a pool of rebate (rewards)
- the CA acts as a manager portal between the DAO, the flow of rebates, and the public. 
- With the public functions on this CA, wallets can vote on the DAO address
- Voting is restricted to once per wallet, voters must carryout a selfless act and nominate an address in order to cast a vote on the next DAO(must be a contract)
- authorized parties will receive a portion of a rare governance coin, at this time each transfer between contract calls costs luck, and lucky number 7 sets the tone. 
-- more features not noted... more intel soon, it's time for dinner 🍕